Refactor event page into standalone component
This will make it possible to show a preview of events in the submission flow using the same UI as the public-facing page.
